Rework the Gensio and Accepter structure
Don't keep a separate pointer to another gensio or accepter in the
structure, now use the same one passed to the user and in the callback.
This required that an Arc<> be returned from the new() functions, but
it was already passed to new_connection() and new_channel() callbacks,
so this is really more consistent.

Signed-off-by: Corey Minyard <minyard@acm.org>